requirements.txt
langgraph>=0.1.0
langchain-core>=0.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
duckduckgo-search>=5.3.0
pydantic>=2.0.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.30.0
hypothesis>=6.100.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
slowapi>=0.1.9
